core/export.py
import uuid
from import_export import resources
from django.utils.timezone import now

class BaseModelResource(resources.ModelResource):
    def before_import_row(self, row, **kwargs):
        print(f"Processing row: {row}")  

        if 'id' not in row or not row['id']:
            row['id'] = str(uuid.uuid4())  
            print(f"Generated new UUID: {row['id']}")
        row.setdefault('active', True)
        row.setdefault('_created_by', None)
        row.setdefault('_updated_by', None)
        row.setdefault('_created_at', now())
        row.setdefault('_updated_at', now())

        print(f"Updated row: {row}") 

        super().before_import_row(row, **kwargs)