requirements.txt
Django>5.0,<5.2
pillow==10.2.0
django-extensions==3.2.3
drf-extensions==0.7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-debug-toolbar==6.1.0

django-cors-headers==4.6.0
django-filter==25.1
cryptography==41.0.7
PyJWT==2.7.0
drf-flex-fields==1.0.2
daphne==4.1.2
psycopg2-binary==2.9.10
openpyxl==3.1.5
django-currentuser==0.8.0
django-render-block==0.10
gunicorn==23.0.0
django-pickling==1.0
django-cleanup==9.0.0
django-simple-history==3.7.0
drf-standardized-errors==0.14.1
channels==4.3.1
channels-redis==4.3.0
gunicorn==23.0.0
uvicorn==0.34.0
django-import-export==4.3.7
dotenv
djoser==2.3.1
logger
pandas==2.2.3
python-dotenv==1.1.0
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
django-eventstream==5.3.1
django-celery-beat==2.8.0

python-json-logger==3.3.0

django-auditlog==3.1.2

flower==2.0.1

drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1

python-jose==3.5.0

persiantools==5.4.0
Werkzeug>=1.0,<4

#----- minIO
django-storages==1.14.6
boto3==1.40.61

#===== logs
websockets==15.0.1
user-agents==2.2.0
 geoip2==5.1.0
#===== click house
clickhouse_connect==0.9.2
